pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ipymini"
version = "0.1.0"
description = "A small, readable, Python-only Jupyter kernel with strong IPython parity."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "ipython>=8.18",
  "jupyter_client>=8.6",
  "jupyter_core>=5.5",
  "pyzmq>=25.1",
]

[project.optional-dependencies]
test = [
  "pytest>=7.4",
  "matplotlib>=3.8",
  "ipykernel>=6.29",
]
dev = [
  "pytest>=7.4",
  "matplotlib>=3.8",
  "ipykernel>=6.29",
]

[project.scripts]
ipymini = "ipymini.__main__:main"

[tool.setuptools]
packages = ["ipymini"]
